requirements.txt
# Core
streamlit>=1.31.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Embeddings & Vector Store
openai>=1.12.0
chromadb>=0.4.22
pinecone-client>=3.0.0
sentence-transformers>=2.3.0

# LangChain
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
tiktoken>=0.5.0

# Utilities
tenacity>=8.2.0
numpy>=1.26.0
pandas>=2.2.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
